backend/requirements.txt
fastapi==0.121.1
uvicorn==0.38.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
textblob==0.19.0
yake==0.6.0
pydantic==2.12.4
starlette==0.49.3

